US607 [Service] Calculate India Other Services - Non-Air Charges

+++ apac-services-exchange-order-module/apac-services-exchange-order-api/src/main/java/com/cwt/bpg/cbt/exchange/order/model/Product.java
@@ -14,7 +14,7 @@ public class Product implements Serializable {
 
 	private String description;
 
-	private Integer gst;
+	private Double gst;
 
 	private String sortKey;
 
@@ -50,11 +50,11 @@ public class Product implements Serializable {
 		this.description = description;
 	}
 
-	public Integer getGst() {
+	public Double getGst() {
 		return gst;
 	}
 
-	public void setGst(Integer gst) {
+	public void setGst(Double gst) {
 		this.gst = gst;
 	}
